Move functions, remove errors
