+++ gpu/src/metal/abstractions/errors.rs
@@ -2,8 +2,6 @@ use thiserror::Error;
 
 #[derive(Debug, Error)]
 pub enum MetalError {
-    #[error("Could not calculate {1} root of unity")]
-    RootOfUnityError(String, u64),
     #[error("Couldn't find a system default device for Metal")]
     MetalDeviceNotFound(),
     #[error("Couldn't create a new Metal library: {0}")]
